Added content to whales-session.tex and new lstenv
Added new content.
Also defined a new lstenvironment called snippet
The snippet environment will not allow pagebreak in the middle of a listing
Also some style to snippet listing, but yadda yadda.

Use \begin{snippet} when defining a snippet on a page.
Only use \begin{lstlisting} when marking up a lot of code.